move interleaved copy/fill tests to module
